+++ .gitignore
@@ -9,3 +9,4 @@ Debug/
 /win/xdagwallet/xdagwallet.vcxproj.user
 /win/ipch
 /win/xdagwallet/xdagwallet.aps
+/win/xdaglib/xdaglib.vcxproj.user
